# docs/swagger.yaml
definitions:
  api.AdClickData:
    properties:
      click_id:
        example: clk_7788992233
        type: string
      jump_url:
        example: https://landing.com?clk=clk_7788992233
        type: string
    type: object
  api.AdClickRequest:
    properties:
      ad_app_id:
        example: app_10001
        type: string
      adjust_adid:
        example: adj_xyz_9988
        type: string
      bid_id:
        example: bid_20260908_0001
        type: string
      creative_id:
        example: cr_90002
        type: string
      timestamp:
        example: 1774213465
        type: integer
      user_id:
        example: user_abc_12345
        type: string
    type: object
  api.AdClickResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/api.AdClickData'
      msg:
        example: success
        type: string
    type: object
  api.AdImpressionRequest:
    properties:
      ad_app_id:
        example: app_10001
        type: string
      bid_id:
        example: bid_20260908_0001
        type: string
      creative_id:
        example: cr_90002
        type: string
      timestamp:
        example: 1774213460
        type: integer
    type: object
  api.AdListData:
    properties:
      ad_list:
        items:
          $ref: '#/definitions/api.AdListItem'
        type: array
      loopable:
        description: |-
          Loopable 整个广告列表是否允许在客户端缓存期内循环播放。
          当且仅当列表内所有素材均可循环（视频）时为 true，任一图片素材出现则为 false。
        example: true
        type: boolean
    type: object
  api.AdListItem:
    properties:
      ad_style:
        description: AdStyle 广告样式列表（单条广告当前的生效样式；为兼容客户端以数组形式接收而使用数组）。
        example:
        - REWARDED_VIDEO
        items:
          type: string
        type: array
      bid_id:
        example: bid_20260908_0001
        type: string
      click_url:
        example: https://landing.com?clk={CLICK_ID}
        type: string
      creative_id:
        example: cr_90002
        type: string
      height:
        example: 1920
        type: integer
      material_type:
        example: VIDEO
        type: string
      material_url:
        example: https://cdn.example.com/creatives/abc123.mp4
        type: string
      required_duration:
        description: |-
          RequiredDuration 客户端须满足的展示时长（秒），按样式取值：
          REWARDED_VIDEO=素材时长（看满才可上报 video-complete 发奖）；
          FEED=素材时长（客户端据此锁定列表滑动，满 N 秒后才可划走，图片素材兜底 3）；
          SPLASH=固定 3；其余样式为 0。
        example: 15
        type: integer
      target_scene:
        example:
        - APP_LAUNCH
        items:
          type: string
        type: array
      width:
        example: 1080
        type: integer
    type: object
  api.AdListRequest:
    properties:
      ad_app_id:
        example: app_10001
        type: string
      adjust_adid:
        example: adj_xyz_9988
        type: string
      count:
        example: 5
        type: integer
      ip:
        example: 192.168.1.100
        type: string
      os:
        example: ios
        type: string
      user_id:
        example: user_abc_12345
        type: string
    type: object
  api.AdListResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/api.AdListData'
      msg:
        example: success
        type: string
    type: object
  api.AdVideoCompleteRequest:
    properties:
      ad_app_id:
        example: app_10001
        type: string
      bid_id:
        example: bid_20260908_0001
        type: string
      creative_id:
        example: cr_90002
        type: string
      timestamp:
        example: 1774213495
        type: integer
      user_id:
        example: user_abc_12345
        type: string
    type: object
  api.ClientSimpleResponse:
    properties:
      code:
        example: 200
        type: integer
      msg:
        example: success
        type: string
    type: object
info:
  contact: {}
  description: 客户端 SDK 对接用接口：批量获取广告、曝光/点击/完播埋点。所有接口通过 HTTP Header `X-Api-Key` 鉴权（值为后台分配的
    App API Key）。
  title: 广告投放中心 客户端接口
  version: "1.0"
paths:
  /v1/ad/click:
    post:
      consumes:
      - application/json
      description: 用户点击广告跳转前调用，换取 click_id（及 jump_url）拼进落地页；服务端按计费方式记一次有效点击。
      parameters:
      - description: App API Key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: 点击埋点请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdClickRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdClickResponse'
      summary: 点击埋点
      tags:
      - 客户端接口
  /v1/ad/impression:
    post:
      consumes:
      - application/json
      description: 素材成功渲染并被用户看到时上报，触发 CPM 扣费与展现统计。
      parameters:
      - description: App API Key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: 曝光埋点请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdImpressionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClientSimpleResponse'
      summary: 曝光埋点
      tags:
      - 客户端接口
  /v1/ad/list:
    post:
      consumes:
      - application/json
      description: 一次请求跨全部广告样式下发，按综合得分汇总取前 count 条；每条素材生成 bid_id（格式 bid_{日期}{时分秒}_{随机}，服务端缓存到
        Redis），供后续曝光/点击/完播埋点闭环归因。
      parameters:
      - description: App API Key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: 批量获取广告请求（body 可空，参数亦可走 query）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdListResponse'
      summary: 批量获取广告
      tags:
      - 客户端接口
  /v1/ad/video-complete:
    post:
      consumes:
      - application/json
      description: 仅 REWARDED_VIDEO 样式：观看达到 required_duration 秒或自然播完时上报，触发 S2S 回调发奖。
      parameters:
      - description: App API Key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: 视频播放完毕请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AdVideoCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ClientSimpleResponse'
      summary: 视频播放完毕
      tags:
      - 客户端接口
schemes:
- https
- http
swagger: "2.0"